Fix GlobalInfoWatcher tests
