+++ app/src/marketplace/utils/web3.js
@@ -3,9 +3,9 @@
 import BN from 'bignumber.js'
 import Web3 from 'web3'
 
-import { StreamrWeb3, getPublicWeb3 } from '$shared/web3/web3Provider'
+import { StreamrWeb3 } from '$shared/web3/web3Provider'
 import getConfig from '$shared/web3/config'
-import type { SmartContractCall, Hash } from '$shared/flowtype/web3-types'
+import type { SmartContractCall } from '$shared/flowtype/web3-types'
 import { getContract, call } from '../utils/smartContract'
 
 import { fromAtto } from './math'
@@ -26,10 +26,3 @@ export const getDataTokenBalance = (web3Instance: StreamrWeb3): SmartContractCal
 
 export const isWeb3Injected = (web3Instance: StreamrWeb3): boolean =>
     web3Instance && (web3Instance.currentProvider != null)
-
-export const hasTransactionCompleted = (txHash: Hash): Promise<boolean> => {
-    const web3 = getPublicWeb3()
-
-    return web3.eth.getTransaction(txHash)
-        .then((trx) => !!trx && trx.blockNumber !== null)
-}
