+++ app/src/shared/test/utils/web3.js
@@ -1,7 +1,9 @@
+import assert from 'assert-diff'
 import sinon from 'sinon'
 
 import * as all from '$shared/utils/web3'
 import * as getConfig from '$shared/web3/config'
+import * as getWeb3 from '$shared/web3/web3Provider'
 
 describe('web3 utils', () => {
     let sandbox
@@ -36,6 +38,54 @@ describe('web3 utils', () => {
             } catch (e) {
                 done()
             }
+        })
+    })
+
+    describe('hasTransactionCompleted', () => {
+        it('returns true if transaction has a block number', async () => {
+            const trx = {
+                blockNumber: 12345,
+            }
+            const transactionStub = sandbox.stub().callsFake(() => Promise.resolve(trx))
+            const publicWeb3Stub = {
+                eth: {
+                    getTransaction: transactionStub,
+                },
+            }
+            sandbox.stub(getWeb3, 'getPublicWeb3').callsFake(() => publicWeb3Stub)
+
+            const result = await all.hasTransactionCompleted('0x123')
+            assert.deepStrictEqual(true, result)
+        })
+
+        it('returns false if transaction doesnt have a block number', async () => {
+            const trx = {
+                blockNumber: null,
+            }
+            const transactionStub = sandbox.stub().callsFake(() => Promise.resolve(trx))
+            const publicWeb3Stub = {
+                eth: {
+                    getTransaction: transactionStub,
+                },
+            }
+            sandbox.stub(getWeb3, 'getPublicWeb3').callsFake(() => publicWeb3Stub)
+
+            const result = await all.hasTransactionCompleted('0x123')
+            assert.deepStrictEqual(false, result)
+        })
+
+        it('returns false if transaction is null', async () => {
+            const trx = null
+            const transactionStub = sandbox.stub().callsFake(() => Promise.resolve(trx))
+            const publicWeb3Stub = {
+                eth: {
+                    getTransaction: transactionStub,
+                },
+            }
+            sandbox.stub(getWeb3, 'getPublicWeb3').callsFake(() => publicWeb3Stub)
+
+            const result = await all.hasTransactionCompleted('0x123')
+            assert.deepStrictEqual(false, result)
         })
     })
 })
