+++ app/src/shared/utils/web3.js
@@ -2,10 +2,11 @@
 
 import { I18n } from 'react-redux-i18n'
 
-import { StreamrWeb3 } from '$shared/web3/web3Provider'
+import { StreamrWeb3, getPublicWeb3 } from '$shared/web3/web3Provider'
 import getConfig from '$shared/web3/config'
 import { ethereumNetworks } from '$shared/utils/constants'
 import { WrongNetworkSelectedError } from '$shared/errors/Web3/index'
+import type { Hash } from '$shared/flowtype/web3-types'
 
 export const checkEthereumNetworkIsCorrect = (web3Instance: StreamrWeb3): Promise<void> => web3Instance.getEthereumNetwork()
     .then((network) => {
@@ -19,3 +20,10 @@ export const checkEthereumNetworkIsCorrect = (web3Instance: StreamrWeb3): Promis
             }))
         }
     })
+
+export const hasTransactionCompleted = (txHash: Hash): Promise<boolean> => {
+    const web3 = getPublicWeb3()
+
+    return web3.eth.getTransaction(txHash)
+        .then((trx) => !!trx && trx.blockNumber !== null)
+}
