+++ app/test/unit/utils/web3.test.js
@@ -61,52 +61,4 @@ describe('web3 utils', () => {
             assert.deepStrictEqual('2209', result)
         })
     })
-
-    describe('hasTransactionCompleted', () => {
-        it('returns true if transaction has a block number', async () => {
-            const trx = {
-                blockNumber: 12345,
-            }
-            const transactionStub = sandbox.stub().callsFake(() => Promise.resolve(trx))
-            const publicWeb3Stub = {
-                eth: {
-                    getTransaction: transactionStub,
-                },
-            }
-            sandbox.stub(getWeb3, 'getPublicWeb3').callsFake(() => publicWeb3Stub)
-
-            const result = await all.hasTransactionCompleted('0x123')
-            assert.deepStrictEqual(true, result)
-        })
-
-        it('returns false if transaction doesnt have a block number', async () => {
-            const trx = {
-                blockNumber: null,
-            }
-            const transactionStub = sandbox.stub().callsFake(() => Promise.resolve(trx))
-            const publicWeb3Stub = {
-                eth: {
-                    getTransaction: transactionStub,
-                },
-            }
-            sandbox.stub(getWeb3, 'getPublicWeb3').callsFake(() => publicWeb3Stub)
-
-            const result = await all.hasTransactionCompleted('0x123')
-            assert.deepStrictEqual(false, result)
-        })
-
-        it('returns false if transaction is null', async () => {
-            const trx = null
-            const transactionStub = sandbox.stub().callsFake(() => Promise.resolve(trx))
-            const publicWeb3Stub = {
-                eth: {
-                    getTransaction: transactionStub,
-                },
-            }
-            sandbox.stub(getWeb3, 'getPublicWeb3').callsFake(() => publicWeb3Stub)
-
-            const result = await all.hasTransactionCompleted('0x123')
-            assert.deepStrictEqual(false, result)
-        })
-    })
 })
